requirements.txt
Flask
Flask-Session
python-dotenv
langchain-google-genai
gunicorn
